vim: dbug: simplify :Dbg command
Now it doesn't receive any parameter and it is only responsible for
starting the gdb python server (that starts the gdb client).

g:dbug_gdb_path can be set for the path of the gdb client (if a custom
one is used). If this variable is not defined the default gdb client is
used.

Signed-off-by: Radu Alexe <alexeRadu2007@gmail.com>