coc.nvim: basic configuration
There are a few drawback for this plugin:
- MSYS2 issues with the fact that it uses native nodejs (which in turn
uses windows-style paths)
- it's dependant of nodejs (a dependency not yet needed)
- it needs other third party software installed (clangd, nodejs, npm)
and therefore will not work out of the box

Yet it seems the be the most easy and complete plugin for completion.
Another interesting option is using neovim which has already merge an
LSP client into it's source code. This probably is the way forward but
for now I'll keep evaluating coc.nvim in a separate branch.

Signed-off-by: Radu Alexe <alexeRadu2007@gmail.com>