nvim: set tabsize to different values depending on the file location
For telescope lua files the author has decided to use 2 spaces for each
tab. So I'm detecting that the file is located inside the neovim data
directory and if so I'm setting the tabsize to 2 spaces.

Further on maybe additional filtering may be necessary since some plugin
authors may have slightly different preferences but for now it will
suffice.

Signed-off-by: Radu Alexe <alexeradu2007@gmail.com>