nvim: change how package configuration is done
The function pkg_config(name, config) checks if it finds a package and
only if it does calls the pkg.setup(config) function.
Therefore if a package is not installed it will not fail.

Additionally each package now gets a file with the exact name in the
lua/packages folder. Each such file should return a .config dictionary
that is passed to the pkg.setup() function.