nvim: create daemon plugin
This plugin uses the plenary Job api to create a job that runs a command
(with arguments) and outputs both the stdout and stderr to a file.
No input is sent to the process.

This is usually used for servers or daemons that need to be
started/stopped from neovim for which we need the output.

As an example a GDB server job could be create using:

```lua
require('daemon').create {
	local cmd  = "<path-to-gdb-server>"
	local args  = {"param1", "param2"}

	require('daemon').create {
	    name    = 'GDB',
	    command = cmd,
	    args    = args,
	}
}
```

In addition the job must be started/stopped with the commands:

```vim
:DaemonStart GDB
:DaemonStop GDB
```

An additional autocommand has been added to stop any Daemon job running
when the editor exits.